+++ libs/integrations/tencent-meeting/exceptions.ts
@@ -1,35 +1,24 @@
 import { HttpException, HttpStatus } from '@nestjs/common';
 
-/**
- * Webhook相关异常基类
- */
+// Webhook exceptions
 export class WebhookException extends HttpException {
   constructor(message: string, status: HttpStatus = HttpStatus.BAD_REQUEST) {
     super(message, status);
   }
 }
 
-/**
- * Webhook签名验证失败异常
- */
 export class WebhookSignatureVerificationException extends WebhookException {
   constructor(platform: string) {
     super(`${platform} Webhook签名验证失败`, HttpStatus.UNAUTHORIZED);
   }
 }
 
-/**
- * Webhook数据解密失败异常
- */
 export class WebhookDecryptionException extends WebhookException {
   constructor(platform: string, error: string) {
     super(`${platform} Webhook数据解密失败: ${error}`, HttpStatus.BAD_REQUEST);
   }
 }
 
-/**
- * Webhook事件类型不支持异常
- */
 export class UnsupportedWebhookEventException extends WebhookException {
   constructor(platform: string, eventType: string) {
     super(
@@ -39,9 +28,6 @@ export class UnsupportedWebhookEventException extends WebhookException {
   }
 }
 
-/**
- * Webhook数据格式异常
- */
 export class WebhookDataFormatException extends WebhookException {
   constructor(platform: string, field: string, expected: string) {
     super(
@@ -51,9 +37,6 @@ export class WebhookDataFormatException extends WebhookException {
   }
 }
 
-/**
- * Webhook处理超时异常
- */
 export class WebhookProcessingTimeoutException extends WebhookException {
   constructor(platform: string, eventType: string, timeout: number) {
     super(
@@ -63,18 +46,12 @@ export class WebhookProcessingTimeoutException extends WebhookException {
   }
 }
 
-/**
- * Webhook重复处理异常
- */
 export class WebhookDuplicateProcessingException extends WebhookException {
   constructor(platform: string, eventId: string) {
     super(`${platform} Webhook事件重复处理: ${eventId}`, HttpStatus.CONFLICT);
   }
 }
 
-/**
- * Webhook配置异常
- */
 export class WebhookConfigException extends WebhookException {
   constructor(platform: string, configKey: string) {
     super(
@@ -84,11 +61,28 @@ export class WebhookConfigException extends WebhookException {
   }
 }
 
-/**
- * Webhook URL验证异常
- */
 export class WebhookUrlVerificationException extends WebhookException {
   constructor(platform: string, error: string) {
     super(`${platform} Webhook URL验证失败: ${error}`, HttpStatus.BAD_REQUEST);
   }
 }
+
+// Platform exceptions
+export class MeetingException extends HttpException {
+  constructor(message: string, status: HttpStatus = HttpStatus.INTERNAL_SERVER_ERROR) {
+    super(message, status);
+  }
+}
+
+export class PlatformConfigException extends MeetingException {
+  constructor(platform: string, configKey: string) {
+    super(`平台配置缺失: ${platform}.${configKey}`, HttpStatus.INTERNAL_SERVER_ERROR);
+  }
+}
+
+export class PlatformApiException extends MeetingException {
+  constructor(platform: string, operation: string, error: string) {
+    super(`${platform} API调用失败 [${operation}]: ${error}`, HttpStatus.BAD_GATEWAY);
+  }
+}
+
