+++ libs/integrations/tencent-meeting/index.ts
@@ -1,0 +1,5 @@
+export * from './tencent.module';
+export * from './tencent-api.service';
+export * from './crypto.util';
+export * from './types';
+export * from './exceptions';
