+++ src/tencent-meeting/controllers/tencent-webhook.controller.ts
@@ -22,13 +22,13 @@ import {
   ApiTencentUrlVerificationDocs,
   ApiTencentEventReceiverDocs,
 } from '../decorators/tencent-webhook.decorators';
-import { verifySignature, aesDecrypt, verifyWebhookUrl } from '../utils/crypto.util';
+import { verifySignature, aesDecrypt, verifyWebhookUrl } from '@libs/integrations/tencent-meeting';
 import { TencentMeetingEvent } from '../types/tencent-webhook-events.types';
 import {
   WebhookSignatureVerificationException,
   WebhookDecryptionException,
   WebhookConfigException,
-} from '../exceptions/webhook.exceptions';
+} from '@libs/integrations/tencent-meeting';
 import { Public } from '@libs/security';
 
 /**
